chore: Refactoring Not Found pages (#26715)
