+++ apps/meteor/client/views/notFound/NotFoundPage.tsx
@@ -1,28 +1,12 @@
-import { Box, States, StatesAction, StatesActions, StatesIcon, StatesSubtitle, StatesTitle } from '@rocket.chat/fuselage';
-import { useRoute, useTranslation } from '@rocket.chat/ui-contexts';
+import { useTranslation } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
 import React from 'react';
 
+import NotFoundState from '../../components/NotFoundState';
+
 const NotFoundPage = (): ReactElement => {
 	const t = useTranslation();
-	const homeRoute = useRoute('home');
-
-	const handleGoHomeClick = (): void => {
-		homeRoute.push();
-	};
-
-	return (
-		<Box display='flex' justifyContent='center' height='full' backgroundColor='surface'>
-			<States>
-				<StatesIcon name='magnifier' />
-				<StatesTitle>{t('Page_not_found')}</StatesTitle>
-				<StatesSubtitle>{t('Page_not_exist_or_not_permission')}</StatesSubtitle>
-				<StatesActions mbs='x16'>
-					<StatesAction onClick={handleGoHomeClick}>{t('Homepage')}</StatesAction>
-				</StatesActions>
-			</States>
-		</Box>
-	);
+	return <NotFoundState title={t('Page_not_found')} subtitle={t('Page_not_exist_or_not_permission')} />;
 };
 
 export default NotFoundPage;
