+++ apps/meteor/client/views/room/RoomNotFound.tsx
@@ -1,20 +1,16 @@
-import { States, StatesIcon, StatesTitle, StatesSubtitle, Box, StatesActions, StatesAction } from '@rocket.chat/fuselage';
+import { Box } from '@rocket.chat/fuselage';
 import { Header as TemplateHeader } from '@rocket.chat/ui-client';
-import { useLayout, useRoute, useTranslation } from '@rocket.chat/ui-contexts';
+import { useLayout, useTranslation } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
 import React from 'react';
 
 import BurgerMenu from '../../components/BurgerMenu';
+import NotFoundState from '../../components/NotFoundState';
 import RoomLayout from './layout/RoomLayout';
 
 const RoomNotFound = (): ReactElement => {
-	const { isMobile } = useLayout();
 	const t = useTranslation();
-	const homeRoute = useRoute('home');
-
-	const handleGoHomeClick = (): void => {
-		homeRoute.push();
-	};
+	const { isMobile } = useLayout();
 
 	return (
 		<RoomLayout
@@ -29,14 +25,7 @@ const RoomNotFound = (): ReactElement => {
 			}
 			body={
 				<Box display='flex' justifyContent='center' height='full'>
-					<States>
-						<StatesIcon name='magnifier' />
-						<StatesTitle>{t('Room_not_found')}</StatesTitle>
-						<StatesSubtitle>{t('Room_not_exist_or_not_permission')}</StatesSubtitle>
-						<StatesActions mbs='x16'>
-							<StatesAction onClick={handleGoHomeClick}>{t('Homepage')}</StatesAction>
-						</StatesActions>
-					</States>
+					<NotFoundState title={t('Room_not_found')} subtitle={t('Room_not_exist_or_not_permission')} />
 				</Box>
 			}
 		/>
